+++ color_analysis_tool/__init__.py
@@ -24,7 +24,7 @@ Basic Usage:
 For more information, visit: https://github.com/MichailSemoglou/color-analysis-tool
 """
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 __author__ = "Michail Semoglou"
 __license__ = "MIT"